# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: CERBERUS API
  description: API definition for the CERBERUS implant-management API.
  version: 2.0.0
# TODO Update with ACG stuff
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
  schemas:
    logindetails:
      properties:
        username:
          type: string
        password:
          type: string
    implant:
      properties:
        _id:
          nullable: false
          type: "string"
        id:
          nullable: false
          type: "string"
        ip:
          nullable: true
          type: "string"
        os:
          nullable: true
          type: "string"
        beaconIntervalSeconds:
          nullable: false
          type: "number"
        lastCheckinTime:
          nullable: false
          type: "string"
        isActive:
          nullable: false
          type: "boolean"
...
paths:
...
  /access/register:
    post: 
      summary: Creates a user, when user management is backed by the database (rather than AD, for example).
      security: []
      tags:
        - Users
        - Access
      requestBody:
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/logindetails"
      responses:
        "200":
          description: User registered successfully 
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "400":
          description: Validation error(s) 
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: Exception thrown, or cannot create users in-app with the configured auth method (eg. Active Directory)
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /access/login:
    post: 
      summary: Logs the user in.
      security: []
      tags:
        - Users
        - Access
      requestBody:
        description: Login credentials
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/logindetails"
      responses:
        "200":
          description: User logged in successfully 
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The bearer token to be used in future authenticated requests.
                  user:
                    type: object
                    properties:
                      id: 
                        type: string
                      name:
                        type: string
                      isAdmin:
                        type: boolean
                  errors:
                    type: array
                    items:
                      type: string
        "401":
          description: Incorrect credentials 
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: Exception thrown, or configured auth method is not supported
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /access/logout/{userId}:
    delete: 
      summary: Logs the user out, by invalidating all bearer tokens they possess.
      security: []
      tags:
        - Users
        - Access
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: Identifier of the user to delete.
      responses:
        "200":
          description: User logged out successfully 
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: Exception thrown
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
  /access/admin:
    put: 
      security:
        - jwt: [Admin]
      summary: Changes admin status of user. Calling user must themselves be an admin.
      tags:
        - Access
      requestBody:
        description: Details of the change to make 
        content:
          application/json:
            schema: 
              type: object
              properties:
                userId: 
                  type: string
                makeAdmin:
                  type: boolean
      responses:
        "200":
          description: User admin status successfully changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "400":
          description: Attempting to change admin status of a non-existent user
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403":
          description: User not logged in or not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: Exception thrown
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
# BEACONS
...
  /implants:
    get:
      security:
        - jwt: [Read]
      summary: Retrieves all implants that the user is authorised to view.
      tags:
        - Implants
      responses:
        "200":
          description: Implant details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  implants:
                    type: array
                    items:
                      $ref: "#/components/schemas/implant"
                  errors:
                    type: array
                    items:
                      type: string
        "403":
          description: User not logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  implants:
                    type: array
                    description: This will be empty
                    items:
                      $ref: "#/components/schemas/implant"
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: An unexpected server-side error
          content:
            application/json:
              schema:
                type: object
                properties:
                  implants:
                    type: array
                    description: This will be empty
                  errors:
                    type: array
                    items:
                      type: string
  /implants/{implantId}:
    delete:
      security:
        - jwt: [Admin]
      summary: Deletes an implant record.
      tags:
        - Implants
      parameters:
        - in: path
          name: implantId
          schema:
            type: string
          required: true
          description: Identifier of the implant to delete.
      responses:
        "200":
          description: Successfully deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "403":
          description: User not logged in, or not an admin.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string
# TASKS
...